Don't exit in package init for version check
It is horrible to have a package exit your application in its init
without the consumer being able to do anything about it.  It is better
to return and error if the lib version is too low and have the consumer
of your package decide what to do instead of exit their application.
This changes it to return a typed error if the lib is to low whenever
you try to create a filter instead of doing this check in init()

Signed-off-by: Michael Crosby <crosbymichael@gmail.com>