all: Bump minimum required version to 2.2.0.
Set the SCMP_FLTATR_CTL_TSYNC attribute in NewFilter, and error out should
that fail.

Since the threads are hidden from the programmer, it is inherently unsafe
to run without this bit set, as the runtime can and does create threads
very early, which are used in subsequent operations (go routines), meaning
that the filter might not be applied, see discussion in #17.

Signed-off-by: Thordur Bjornsson <thorduri@secnorth.net>