golang: Add ActLog test
Add a new test that defaults to ActErrno, allows the syscalls needed to
carry out a basic test, and sets the getpid() syscall to ActLog. The
getpid() syscall is called before the filter is loaded and once again
after the filter is loaded. The test is successful if the return values
match.

The test is skipped when libseccomp is not new enough to support API
level operations or the API level is less than 3.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>