all: various "make lint" fixes
The "*Api" to "*API" function renames might be slightly annoying for
callers, but we haven't hit the all-important v1.0.0 so I think we
can still reserve the right to tweak the API slightly.  Also, it is
unlikely that these API functions are in heavy use by callers at the
moment.

Signed-off-by: Paul Moore <paul@paul-moore.com>