all: fix SCMP_FLTATR_CTL_SSB with old libseccomp
As commit "golang: Add support for SCMP_FLTATR_CTL_SSB"
(276b0ea443de00461570ec512734ec0c3bc6cc42) says, that was added in
libseccomp 2.5.0. However, the elif clause added there meant that if we
are using a libseccomp < 2.4 the define will not be done.

This was not an issue because we weren't testing with such an old
libseccomp, but now we will test against < 2.4 libseccomp in future
patches.

This patch just fixes it by using different if clause for the different
defines, instead of elif.

Signed-off-by: Rodrigo Campos <rodrigo@kinvolk.io>
Acked-by: Tom Hromatka <tom.hromatka@oracle.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>