tests: add missing APILevelIsSupported() guard
Commit "golang: Add support for SCMP_FLTATR_CTL_SSB"
(276b0ea443de00461570ec512734ec0c3bc6cc42) added a call to GetAPI(), and
is not guarded with "APILevelIsSupported()" to only report errors when
it is supported. The rest of the added code had that, it is just missing
for the GetAPI() part.

This patch just adds a "APILevelIsSupported()" guard, as done in the
rest of GetAPI() calls, to make sure it runs correctly with old
libseccomp versions.

Signed-off-by: Rodrigo Campos <rodrigo@kinvolk.io>
Acked-by: Tom Hromatka <tom.hromatka@oracle.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>