all: unify libseccomp version / API level checks
This unifies all the code that is used to check for minimally required
libseccomp version and API level and generate an error.

For errors, VersionError is reused, which is now amended with API level.

For checks, we now have checkVersion and checkAPI, which generate errors
that can be returned as is.

This removes some repetitive code and unifies error messages. As a
side benefit, a user can now check if an error was caused by an old
libseccomp and/or low API level, by using errors.As or checking if the
type is VersionError.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>