Makefile: use golangci-lint for lint target
golint is no more, so remove it.

golangci-lint is a tool that runs multuple linters in parallel.
Default configuration includes go vet (which is being replaced),
errorcheck, and a few other linters (see golangci-lint linters).

It is assumed that users do have golangci-lint installed, or can
figure out how to do that.

Now, since golangci-lint runs govet, it makes sense to modify
the check target to call lint instead of go vet.

README is amended accordingly.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>
Acked-by: Paul Moore <paul@paul-moore.com>
Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>