CI/GHA: test against libseccomp 2.5.2 and HEAD
Instead of testing against whatever version of libseccomp the distro
(Ubuntu 20.04 in this case) provides, do compile and test against the
latest released version, 2.5.2.

While at it, it's relatively easy to enable testing against git tip as
well, so implement it as well. All we have to do is to change the
version from 0.0.0 to some value that is >= current version.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>
Acked-by: Tom Hromatka <tom.hromatka@oracle.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>