{README,CONTRIBUTING}.md: dedup
README.md contains some information about testing the library, also
there are some notes about it in CONTRIBUTING.md. It seems the relevant
text in README better reflects the reality ("make check" performs both
linting and testing, it is sufficient, and it requires golangci-lint),
so let's use it and drop the other.

In other words:
 - replace the testing note in CONTRIBUTING with the ones from README;
 - in README, add a reference to CONTRIBUTING.

[v2: fix a typo that comes from README]

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>