TestFilterAttributeGettersAndSetters: fixup
Commit 23edf066d25 moved the SetBadArchAction(ActInvalid) check
to the end of TestFilterAttributeGettersAndSetters. A few subsequent
commits added more code before it, until finally commit 541420d6f0c
added a conditional t.Skipf before it.

As a result, this check is not performed with latest libseccomp.

Move it back to where it belongs (before the API/version checks).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>