*: gofumpt
This is a result of running "gofumpt -w *.go" on the source code.

gofumpt[1] is a tool similar to (and based on) gofmt, which adds some
some extra rules for better code formatting. Here, it removes some extra
vertical whitespace, and replaces a var() block with a single element
inside with a mere var declaration. Both make sense.

Enable gofumpt in golangci-lint config to make sure future commits are
gofumpt-ed.

[1] https://github.com/mvdan/gofumpt

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>
Acked-by: Paul Moore <paul@paul-moore.com>
Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>