seccomp_internal: use errors.New, add errDefAction
The latter is to be used by the next commit.

While at it, slightly improve the errBadFilter doc string.

NOTE that there are more places that need s/fmt.Errorf/errors.New/, but
as such change is trivial and it will break other commit pending review,
I am going to implement it separately later.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>