TestRuleAddAndLoad: fix for ppc
As reported in [1], the test fails on ppc64le since getpid(2) is
not supposed to ever return an error, and somehow glibc relies on
that assumption, and returns a positive value of error set by
the seccomp rule from its getpid(2) wrapper. This makes the test fail.

While seccomp_rule_add(3) man page was amended to note that behavior
in [2], the test case here was never fixed so it always fails on ppc.

Fix it by replacing getpid(2) with close(2). Make some other cosmetic
changes while at it.

The test case was tested to fail (when the corresponding FilterAddRule
call is commented out):

> seccomp_test.go:647: Syscall listen: want no space left on device, got socket operation on non-socket

[1] https://github.com/seccomp/libseccomp-golang/issues/61
[2] https://github.com/seccomp/libseccomp/pull/333

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>
Acked-by: Tom Hromatka <tom.hromatka@oracle.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>