go.mod: use go 1.19
The meaning of this setting is somewhat vague in the official Go docs.
To my best understanding, this value mostly influences how Go toolchain
deals with go.mod and go.sum files.

In our case, we don't use any other go modules, so go.mod and go.sum are
quite shallow, yet I guess it makes sense to have something sensible in
here.

Switch from 1.14 to 1.19 as at the moment this is the oldest Go version
that is still supported.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>